GovernanceVotingAgent/src/GovernanceVotingAgent/agent.py

from autogen import AssistantAgent
from agoric_sdk import ContractAPI

class GovernanceVotingAgent:
    def __init__(self):
        self.assistant = AssistantAgent("voting_assistant")
        self.contract_api = ContractAPI("https://api.agoric.net")

    def start_voting(self, proposal):
        try:
            # Start the voting process for a proposal
            result = self.contract_api.start_voting(proposal)
            return result
        except Exception as e:
            return f"Failed to start voting: {str(e)}"

    def collect_votes(self, voting_id):
        try:
            # Collect votes for a given voting ID
            votes = self.contract_api.get_votes(voting_id)
            analysis = self.analyze_votes(votes)
            return analysis
        except Exception as e:
            return f"Failed to collect votes: {str(e)}"

    def analyze_votes(self, votes):
        # Implement vote analysis logic here
        total_votes = len(votes)
        yes_votes = sum(1 for vote in votes if vote == "yes")
        no_votes = total_votes - yes_votes
        return {"total": total_votes, "yes": yes_votes, "no": no_votes}

    def announce_results(self, voting_id):
        try:
            # Announce the voting results
            results = self.collect_votes(voting_id)
            announcement = f"Voting ID: {voting_id}, Total Votes: {results['total']}, Yes: {results['yes']}, No: {results['no']}"
            response = self.assistant.handle_input(announcement)
            return response
        except Exception as e:
            return f"Failed to announce results: {str(e)}"

if __name__ == "__main__":
    agent = GovernanceVotingAgent()
    proposal = "Should we implement feature X?"
    start_result = agent.start_voting(proposal)
    print(start_result)
    voting_id = "example_voting_id"
    results = agent.announce_results(voting_id)
    print(results)
